refactor: 할인 정책 프로토콜 구현
