+++ objects/objects/chapter02-movie/DiscountPolicy.swift
@@ -8,23 +8,8 @@
 import Foundation
 
 /**
- 할인 정책
+ 할인 정책 프로토콜
  */
 protocol DiscountPolicy {
-    var conditions: [DiscountCondition] { get set }
-
     func calculateDiscountAmount(screening: Screening) -> Money
-
-    func getDiscountAmount(screening: Screening) -> Money
-}
-
-extension DiscountPolicy {
-    func calculateDiscountAmount(screening: Screening) -> Money {
-        for each in conditions {
-            if each.isSatisfiedBy(screening) {
-                return getDiscountAmount(screening: screening)
-            }
-        }
-        return Money.zero
-    }
 }
